Add graphql service example
This service uses the java graphql api directly, in combination with
javalin for the webserver routing. I'm using the dgs library to convert
the graphql schema into beans that can be consumed by the graphql api,
in the past I've managed those beans directly, and the generation in the
library makes it silghtly easier.

A more detailed README to come.